Getting data from database using Eloquent Model and displaying on view

+++ database/migrations/2020_03_11_212131_create_games_table.php
@@ -17,7 +17,7 @@ class CreateGamesTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('type');
-            $table->float('price', 3, 2);
+            $table->integer('price');
             $table->timestamps();
         });
     }
